requirements.txt
torch==1.11.0
Pillow==9.1.0
torchvision==0.12.0
Flask==2.1.1
flask_cors==3.0.10
python-dotenv==0.20.0
